platform/x86: lenovo-wmi-helpers: Move gamezone enums to wmi-helpers
commit 7e27896e16a1c450085c3fe020eeb1b223880f37 upstream.

In a later patch in the series the thermal mode enum will be accessed
across three separate drivers (wmi-capdata, wmi-gamezonem and wmi-other).
An additional patch in the series will also add a function prototype that
needs to reference this enum in wmi-helpers.h. To avoid having all these
drivers begin to import each others headers, and to avoid declaring an
opaque enum to hande the second case, move the thermal mode enum to
helpers where it can be safely accessed by everything that needs it from
a single import.

While at it, since the gamezone_events_type enum is the only remaining
item in the header, move that as well and remove the gamezone header
entirely.

Cc: stable@vger.kernel.org
Tested-by: Rong Zhang <i@rong.moe>
Link: https://patch.msgid.link/20260510042546.436874-8-derekjohn.clark@gmail.com
(cherry picked from commit b6c0f545c8f94b447e554e57b66d28b163b48837)
